requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.7.2
motor==3.4.0
web3==6.20.0
pydantic==2.7.1
py-solc-x==2.0.2
python-multipart==0.0.9
slowapi==0.1.9
starlette==0.37.2
python-dotenv==1.0.1